Make DiagonalEMFit class
