+++ src/mlpack/methods/gmm/CMakeLists.txt
@@ -9,6 +9,8 @@ set(SOURCES
   diagonal_gmm_impl.hpp
   em_fit.hpp
   em_fit_impl.hpp
+  diagonal_em_fit.hpp
+  diagonal_em_fit_impl.hpp
   no_constraint.hpp
   positive_definite_constraint.hpp
   diagonal_constraint.hpp
